Rename `pin` module to `digital` to match embedded-hal
