+++ src/eh0.rs
@@ -9,8 +9,8 @@ pub use error::MockError;
 
 pub mod adc;
 pub mod delay;
+pub mod digital;
 pub mod i2c;
-pub mod pin;
 pub mod serial;
 pub mod spi;
 #[cfg(feature = "embedded-time")]
