+++ src/eh1.rs
@@ -8,8 +8,8 @@ mod error;
 pub use crate::eh1::error::MockError;
 
 pub mod delay;
+pub mod digital;
 pub mod i2c;
-pub mod pin;
 pub mod pwm;
 pub mod serial;
 pub mod spi;
